Fixed threading. Windows really hates creating many threads very fast, so I moved them to be created at program start. Now we are getting really good FPS, 3 cores gets me ~100FPS (50 previously) on my laptop, but it heats up quite a bit.

+++ Source/Client/Client.cpp
@@ -445,7 +445,7 @@ void Client::play()
 								sf::Sprite cursor; cursor.setTexture(cursorTex);
 								cursor.setOrigin(cursorTex.getSize().x / 2.0f, cursorTex.getSize().y / 2.0f);
 								cursor.setPosition(renderWidth * screenScale * 0.5f, topUiHeight + player->aimElevation * renderHeight * screenScale);
-								cursor.setScale(screenScale, screenScale);
+								cursor.setScale((float)screenScale, (float)screenScale);
 
 								win->draw(cursor);
 							}
